Change from strings to named types
Named types give a lot more assurance into what is actually happening,
so it would be better to move to named types as soon as possible.

Not completely finished, there are tons of commented out methods and
tests.

dnaSequence is now representing a chain of nucleotides of any length.

Kmer represents all mers of length k within a sequence.

A nucleotide is a single protein with it's base letter.

pattens still has all the legacy untyped functions.