Fix GenerateNeighbors function
Changes the generate neighbors function to using dnaSequences instead of
kmers. Kmers are kind of inflexible, and hold pretty much all the same
information as a slice of dnaSequences. The only thing that a kmer gives
you over a slice of dnaSequences is that a kmer will have dnaSequences
with all the same length, but until I find that being useful, I think
all of these function will be using dnaSequences.