Consolidate python code into a single place
Groups together are rosalind projects and other projects so they each
have their own file, but they share the same utils.

Having the rosalind problems in one file that are all stored in a
dictionary so at runtime the desired program can be selected and run.
This will keep the number of new files low since rosalind has a bunch of
tiny problems.