Add hamming distance to rosalind functions
The hamming distance function was already being used a ton in the other
course, so I just made a function mapping in the rosalind python file
that accepted a file with two lines.

Each line in this file has newlines stripped and then the list is
returned to be used by whatever.